feat: move independent data services to packages
