+++ app/src/main/java/DuGraphics/services/data/common/INodeListener.java
@@ -1,4 +1,4 @@
-package DuGraphics.services.data;
+package DuGraphics.services.data.common;
 
 public interface INodeListener<T> {
     void action(Node<T> node);
